refactor(evm): move policy-independent intrinsic-gas helpers off IGasPolicy
The six tx+spec-only helpers (CalculateTokensInCallData, AccessListCost,
AuthorizationListCost, CalculateFloorTokensInAccessList/InCallData, CalculateFloorCost)
lived on IGasPolicy as static members, forcing the self-referential
IGasPolicy<EthereumGasPolicy>.X call shape and bloating the interface. They take no
TSelf and belong with intrinsic-gas calculation. Move them into IntrinsicGasCalculator
(internal static), where AccessListCost was already wrapping them, and call them directly
from EthereumGasPolicy.CalculateIntrinsicGas. Behavior-identical (pure relocation).
Verified: IntrinsicGasCalculator + Eip8037 tests 129/129, build clean.
