+++ src/Nethermind/Nethermind.Evm/GasPolicy/IGasPolicy.cs
@@ -2,11 +2,7 @@
 // SPDX-License-Identifier: LGPL-3.0-only
 
 using System;
-using System.Diagnostics;
-using System.Diagnostics.CodeAnalysis;
-using System.IO;
 using Nethermind.Core;
-using Nethermind.Core.Eip2930;
 using Nethermind.Core.Extensions;
 using Nethermind.Core.Specs;
 using Nethermind.Evm.Precompiles;
@@ -294,75 +290,6 @@ public interface IGasPolicy<TSelf> where TSelf : struct, IGasPolicy<TSelf>
 
     static abstract void OnBeforeInstructionTrace(in TSelf gas, int pc, Instruction instruction, int depth);
     static abstract void OnAfterInstructionTrace(in TSelf gas);
-
-    protected static ulong CalculateTokensInCallData(Transaction transaction, IReleaseSpec spec)
-    {
-        ReadOnlySpan<byte> data = transaction.Data.Span;
-        int totalZeros = data.CountZeros();
-        return (ulong)totalZeros + (ulong)(data.Length - totalZeros) * spec.GasCosts.TxDataNonZeroMultiplier;
-    }
-
-    // 0 when floor pricing is not active.
-    public static ulong CalculateFloorTokensInAccessList(Transaction transaction, IReleaseSpec spec) =>
-        spec.IsEip7981Enabled && transaction.AccessList is { Count: (int addressesCount, int storageKeysCount) }
-            ? (ulong)(addressesCount * Address.Size + storageKeysCount * AccessList.StorageKeySize) * spec.GasCosts.TxDataNonZeroMultiplier
-            : 0;
-
-    public static ulong AccessListCost(Transaction transaction, IReleaseSpec spec, ulong floorTokensInAccessList)
-    {
-        AccessList? accessList = transaction.AccessList;
-        if (accessList is null) return 0;
-
-        if (!spec.UseTxAccessLists)
-        {
-            ThrowInvalidDataException(spec);
-        }
-
-        (int addressesCount, int storageKeysCount) = accessList.Count;
-        return (ulong)addressesCount * GasCostOf.AccessAccountListEntry
-            + (ulong)storageKeysCount * GasCostOf.AccessStorageListEntry
-            + spec.GasCosts.TotalCostFloorPerToken * floorTokensInAccessList;
-
-        [DoesNotReturn, StackTraceHidden]
-        static void ThrowInvalidDataException(IReleaseSpec spec) =>
-            throw new InvalidDataException($"Transaction with an access list received within the context of {spec.Name}. EIP-2930 is not enabled.");
-    }
-
-    public static (ulong RegularCost, ulong StateCost) AuthorizationListCost(Transaction transaction, IReleaseSpec spec)
-    {
-        AuthorizationTuple[]? authList = transaction.AuthorizationList;
-        if (authList is null)
-        {
-            return (0, 0);
-        }
-
-        if (!spec.IsAuthorizationListEnabled)
-        {
-            ThrowAuthorizationListNotEnabled(spec);
-        }
-
-        ulong authCount = (ulong)authList.Length;
-        return spec.IsEip8037Enabled
-            ? (
-                authCount * GasCostOf.PerAuthBaseRegular,
-                authCount * (GasCostOf.NewAccountState + GasCostOf.PerAuthBaseState)
-            )
-            : (authCount * GasCostOf.NewAccount, 0);
-
-        [DoesNotReturn, StackTraceHidden]
-        static void ThrowAuthorizationListNotEnabled(IReleaseSpec releaseSpec) =>
-            throw new InvalidDataException($"Transaction with an authorization list received within the context of {releaseSpec.Name}. EIP-7702 is not enabled.");
-    }
-
-    private static ulong CalculateFloorTokensInCallData(Transaction transaction, IReleaseSpec spec) =>
-        (ulong)transaction.Data.Length * spec.GasCosts.TxDataNonZeroMultiplier;
-
-    protected static ulong CalculateFloorCost(Transaction transaction, IReleaseSpec spec, ulong tokensInCallData, ulong floorTokensInAccessList) => spec switch
-    {
-        { IsEip7976Enabled: true } => GasCostOf.Transaction + (CalculateFloorTokensInCallData(transaction, spec) + floorTokensInAccessList) * spec.GasCosts.TotalCostFloorPerToken,
-        { IsEip7623Enabled: true } => GasCostOf.Transaction + tokensInCallData * spec.GasCosts.TotalCostFloorPerToken,
-        _ => 0
-    };
 }
 
 public readonly record struct IntrinsicGas<TGasPolicy>(TGasPolicy Standard, TGasPolicy FloorGas)
